move codes to create.go
